Updated frontend files
